feat(combobox-web): implement BaseDatasourceOptionsProvider class
